LayoutTests/fast/canvas/webgl/webgl-specific-expected.txt

Tests the a few differences between WebGL and GLES2

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS getError was expected value: INVALID_OPERATION : constant color and constant alpha cannot be used together as source and destination factors in the blend function
PASS getError was expected value: INVALID_OPERATION : constant color and constant alpha cannot be used together as source and destination factors in the blend function
PASS getError was expected value: INVALID_OPERATION : constant color and constant alpha cannot be used together as source and destination factors in the blend function
PASS getError was expected value: INVALID_OPERATION : constant color and constant alpha cannot be used together as source and destination factors in the blend function
PASS getError was expected value: INVALID_OPERATION : constant color and constant alpha cannot be used together as source and destination factors in the blend function
PASS getError was expected value: INVALID_OPERATION : constant color and constant alpha cannot be used together as source and destination factors in the blend function
PASS getError was expected value: INVALID_OPERATION : constant color and constant alpha cannot be used together as source and destination factors in the blend function
PASS getError was expected value: INVALID_OPERATION : constant color and constant alpha cannot be used together as source and destination factors in the blend function
PASS getError was expected value: INVALID_OPERATION : constant color and constant alpha cannot be used together as source and destination factors in the blend function
PASS getError was expected value: INVALID_OPERATION : constant color and constant alpha cannot be used together as source and destination factors in the blend function
PASS getError was expected value: INVALID_OPERATION : constant color and constant alpha cannot be used together as source and destination factors in the blend function
PASS getError was expected value: INVALID_OPERATION : constant color and constant alpha cannot be used together as source and destination factors in the blend function
PASS getError was expected value: INVALID_OPERATION : constant color and constant alpha cannot be used together as source and destination factors in the blend function
PASS getError was expected value: INVALID_OPERATION : constant color and constant alpha cannot be used together as source and destination factors in the blend function
PASS getError was expected value: INVALID_OPERATION : constant color and constant alpha cannot be used together as source and destination factors in the blend function
PASS getError was expected value: INVALID_OPERATION : constant color and constant alpha cannot be used together as source and destination factors in the blend function
PASS getError was expected value: INVALID_OPERATION : depthRange should generate INVALID_OPERATION if zNear is greater than zFar
PASS getError was expected value: NO_ERROR : stencilMask should generate no error
PASS getError was expected value: INVALID_OPERATION : a different mask value for front and back facing is illegal
PASS getError was expected value: INVALID_OPERATION : a different mask value for front and back facing is illegal
PASS getError was expected value: NO_ERROR : stencilFunc should generate no error
PASS getError was expected value: INVALID_OPERATION : a different reference value for front and back facing is illegal
PASS getError was expected value: INVALID_OPERATION : a different reference value for front and back facing is illegal
PASS getError was expected value: INVALID_OPERATION : a different mask value for front and back facing is illegal
PASS getError was expected value: INVALID_OPERATION : a different mask value for front and back facing is illegal
PASS undefined is undefined.
PASS undefined is undefined.
PASS gl.getParameter(gl.UNPACK_COLORSPACE_CONVERSION_WEBGL) is gl.BROWSER_DEFAULT_WEBGL
PASS gl.getParameter(gl.UNPACK_COLORSPACE_CONVERSION_WEBGL) is gl.NONE
PASS getError was expected value: NO_ERROR : set/get UNPACK_COLORSPACE_CONVERSION_WEBGL should generate no error
PASS successfullyParsed is true

TEST COMPLETE
